requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.5
django-cors-headers==3.7.0
django-waitress==0.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.18.2
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.26.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
jdcal==1.4.1
mypy-extensions==0.4.3
numpy==1.20.1
oauthlib==3.1.0
openpyxl==3.0.6
packaging==20.9
pandas==1.2.1
pathspec==0.8.1
pefile==2019.4.18
Pillow==8.1.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyinstaller-hooks-contrib==2020.11
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
pywin32-ctypes==0.2.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
waitress==1.4.4
wfastcgi==3.0.0
whitenoise==5.2.0
